Split user information to its own package
This package will be used to add/remove users from local groups as this
information is not part of the cache.
Adapt imports and tests for the new struct location.
