+++ internal/cache/db.go
@@ -63,23 +63,6 @@ type Cache struct {
 	cleanupQuitted chan struct{}
 }
 
-// UserInfo is the user information returned by the broker. We use that to build our own buckets content.
-type UserInfo struct {
-	Name  string
-	UID   int
-	Gecos string
-	Dir   string
-	Shell string
-
-	Groups []GroupInfo
-}
-
-// GroupInfo is the group information returned by the broker. We use that to build our own buckets content.
-type GroupInfo struct {
-	Name string
-	GID  *int
-}
-
 // userDB is the struct stored in json format in the bucket.
 type userDB struct {
 	UserPasswdShadow
