+++ internal/services/pam/pam.go
@@ -11,6 +11,7 @@ import (
 	"github.com/ubuntu/authd/internal/brokers/responses"
 	"github.com/ubuntu/authd/internal/cache"
 	"github.com/ubuntu/authd/internal/log"
+	"github.com/ubuntu/authd/internal/users"
 	"github.com/ubuntu/decorate"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -194,7 +195,7 @@ func (s Service) IsAuthenticated(ctx context.Context, req *authd.IARequest) (res
 
 	// Update database on granted auth.
 	if access == responses.AuthGranted {
-		var user cache.UserInfo
+		var user users.UserInfo
 		if err := json.Unmarshal([]byte(data), &user); err != nil {
 			return nil, fmt.Errorf("user data from broker invalid: %v", err)
 		}
