+++ internal/services/pam/testdata/TestIsAuthenticated/golden/error_when_broker_returns_invalid_userinfo/IsAuthenticated
@@ -1,4 +1,4 @@
 FIRST CALL:
 	access: 
 	msg: 
-	err: rpc error: code = Unknown desc = can't check authentication: invalid user information provided by the broker ("not valid"): message is not JSON formatted: json: cannot unmarshal string into Go value of type struct { cache.UserInfo; UUID string; UGID string; Groups []struct { Name string; UGID string } }
+	err: rpc error: code = Unknown desc = can't check authentication: invalid user information provided by the broker ("not valid"): message is not JSON formatted: json: cannot unmarshal string into Go value of type struct { users.UserInfo; UUID string; UGID string; Groups []struct { Name string; UGID string } }
